requirements.txt
fastapi~=0.103.1
SQLAlchemy~=1.4.49
pydantic~=2.3.0
requests~=2.31.0
python-dotenv~=1.0.0
pyfunctional==1.4.3
mysqlclient==2.2.0
loguru~=0.7.2